rename weapon arg
