+++ Assets/Sc_Obj/Weapons/WeaponArg.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: d8762b7003e19224dbf0344b533097c6
+timeCreated: 1564994247
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
