--- Assets/Sc_Obj/Weapons/Weapon_Arg.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: e76b5bf74ca47334b863b8a68291dd66
-timeCreated: 1564994247
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
